put everything in coalignment
